Add DropService for loot drop replacement logic and refactor PlayerDropService
