+++ surf-radnom-drops-event/src/main/kotlin/dev/slne/surf/event/randomdrops/listener/random/RandomLootTable.kt
@@ -1,6 +1,6 @@
 package dev.slne.surf.event.randomdrops.listener.random
 
-import dev.slne.surf.event.randomdrops.service.PlayerDropService
+import dev.slne.surf.event.randomdrops.service.DropService
 import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
 import org.bukkit.event.Listener
@@ -9,6 +9,6 @@ import org.bukkit.event.world.LootGenerateEvent
 object RandomLootTable : Listener {
     @EventHandler(priority = EventPriority.LOWEST)
     fun onLootGenerate(event: LootGenerateEvent) {
-        event.setLoot(PlayerDropService.generateReplacedLootDrop(event.loot))
+        event.setLoot(DropService.generateReplacedLootDrop(event.loot))
     }
 }
