+++ surf-radnom-drops-event/src/main/kotlin/dev/slne/surf/event/randomdrops/service/PlayerDropService.kt
@@ -1,12 +1,7 @@
 package dev.slne.surf.event.randomdrops.service
 
 import dev.slne.surf.event.randomdrops.data.PlayerDataStorage
-import dev.slne.surf.event.randomdrops.random.RandomDropSelector
 import dev.slne.surf.surfapi.bukkit.api.nms.bridges.nmsLootTableBridge
-import dev.slne.surf.surfapi.core.api.util.mutableObject2ObjectMapOf
-import dev.slne.surf.surfapi.core.api.util.mutableObjectListOf
-import it.unimi.dsi.fastutil.objects.ObjectList
-import net.kyori.adventure.key.Key
 import org.bukkit.Registry
 import org.bukkit.damage.DamageSource
 import org.bukkit.entity.LivingEntity
@@ -33,23 +28,4 @@ object PlayerDropService {
         return nmsLootTableBridge.getDifferentLootTable(entity, damageSource, replacedType, true)
     }
 
-    fun generateReplacedLootDrop(original: Collection<ItemStack>): ObjectList<ItemStack> {
-        val replacementKeys = mutableObject2ObjectMapOf<Key, Key>()
-        val replacements = mutableObjectListOf<ItemStack>()
-
-        for (stack in original) {
-            val originalType = stack.type.asItemType()
-            if (originalType == null) {
-                replacements.add(stack)
-                continue
-            }
-            val replacementKey = replacementKeys.computeIfAbsent(originalType.key) {
-                RandomDropSelector.selectRandomBlockDrop(originalType.itemRarity).key
-            }
-            val replacementType = Registry.ITEM.getOrThrow(replacementKey)
-            replacements.add(replacementType.createItemStack(stack.amount))
-        }
-
-        return replacements
-    }
 }
